requirements.txt
google-generativeai
flask
streamlit